Add ability to mirror database writes
- Adds a `mirrorable` option to `Stockpile.redis` in addition to
yielding the db for the key you passed (`default`) it will also yield
for a matching key if it exists (`default_2x`). This allows you to
execute a single Redis command inside of Stockpile against two databases.
- Updates Stockpile so that locks, cache writes, and lock expirations
use this new `Stockpile.redis_2x` method, leaving the read methods
alone so they continue to call execute once

The idea would be we could deploy this per Stockpile database, wait
for the longest TTL on that database, then deploy once more  promoting
the `default_2x` database key to be the `default` database key so
reads now take place on the new cluster.